+++ src/Service/CloudFormation/src/CloudFormationClient.php
@@ -88,104 +88,71 @@ class CloudFormationClient extends AbstractApi
             case 'us-west-1':
             case 'us-west-2':
                 return [
-                    'endpoint' => 'https://cloudformation.%region%.amazonaws.com',
+                    'endpoint' => "https://cloudformation.$region.amazonaws.com",
                     'signRegion' => $region,
                     'signService' => 'cloudformation',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'cn-north-1':
             case 'cn-northwest-1':
                 return [
-                    'endpoint' => 'https://cloudformation.%region%.amazonaws.com.cn',
+                    'endpoint' => "https://cloudformation.$region.amazonaws.com.cn",
                     'signRegion' => $region,
                     'signService' => 'cloudformation',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'us-gov-east-1':
             case 'us-gov-west-1':
                 return [
-                    'endpoint' => 'https://cloudformation.%region%.amazonaws.com',
+                    'endpoint' => "https://cloudformation.$region.amazonaws.com",
                     'signRegion' => $region,
                     'signService' => 'cloudformation',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'us-iso-east-1':
                 return [
-                    'endpoint' => 'https://cloudformation.%region%.c2s.ic.gov',
+                    'endpoint' => "https://cloudformation.$region.c2s.ic.gov",
                     'signRegion' => $region,
                     'signService' => 'cloudformation',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'us-isob-east-1':
                 return [
-                    'endpoint' => 'https://cloudformation.%region%.sc2s.sgov.gov',
+                    'endpoint' => "https://cloudformation.$region.sc2s.sgov.gov",
                     'signRegion' => $region,
                     'signService' => 'cloudformation',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'us-east-1-fips':
                 return [
                     'endpoint' => 'https://cloudformation-fips.us-east-1.amazonaws.com',
                     'signRegion' => 'us-east-1',
                     'signService' => 'cloudformation',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'us-east-2-fips':
                 return [
                     'endpoint' => 'https://cloudformation-fips.us-east-2.amazonaws.com',
                     'signRegion' => 'us-east-2',
                     'signService' => 'cloudformation',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'us-west-1-fips':
                 return [
                     'endpoint' => 'https://cloudformation-fips.us-west-1.amazonaws.com',
                     'signRegion' => 'us-west-1',
                     'signService' => 'cloudformation',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'us-west-2-fips':
                 return [
                     'endpoint' => 'https://cloudformation-fips.us-west-2.amazonaws.com',
                     'signRegion' => 'us-west-2',
                     'signService' => 'cloudformation',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
         }
 
         throw new UnsupportedRegion(sprintf('The region "%s" is not supported by "CloudFormation".', $region));
-    }
-
-    protected function getServiceCode(): string
-    {
-        return 'cloudformation';
-    }
-
-    protected function getSignatureScopeName(): string
-    {
-        return 'cloudformation';
-    }
-
-    protected function getSignatureVersion(): string
-    {
-        return 'v4';
     }
 }