+++ src/Service/CodeDeploy/src/CodeDeployClient.php
@@ -60,113 +60,78 @@ class CodeDeployClient extends AbstractApi
             case 'us-west-1':
             case 'us-west-2':
                 return [
-                    'endpoint' => 'https://codedeploy.%region%.amazonaws.com',
+                    'endpoint' => "https://codedeploy.$region.amazonaws.com",
                     'signRegion' => $region,
                     'signService' => 'codedeploy',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'cn-north-1':
             case 'cn-northwest-1':
                 return [
-                    'endpoint' => 'https://codedeploy.%region%.amazonaws.com.cn',
+                    'endpoint' => "https://codedeploy.$region.amazonaws.com.cn",
                     'signRegion' => $region,
                     'signService' => 'codedeploy',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'us-gov-east-1':
             case 'us-gov-west-1':
                 return [
-                    'endpoint' => 'https://codedeploy.%region%.amazonaws.com',
+                    'endpoint' => "https://codedeploy.$region.amazonaws.com",
                     'signRegion' => $region,
                     'signService' => 'codedeploy',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'us-iso-east-1':
                 return [
-                    'endpoint' => 'https://codedeploy.%region%.c2s.ic.gov',
+                    'endpoint' => "https://codedeploy.$region.c2s.ic.gov",
                     'signRegion' => $region,
                     'signService' => 'codedeploy',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'us-east-1-fips':
                 return [
                     'endpoint' => 'https://codedeploy-fips.us-east-1.amazonaws.com',
                     'signRegion' => 'us-east-1',
                     'signService' => 'codedeploy',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'us-east-2-fips':
                 return [
                     'endpoint' => 'https://codedeploy-fips.us-east-2.amazonaws.com',
                     'signRegion' => 'us-east-2',
                     'signService' => 'codedeploy',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'us-gov-east-1-fips':
                 return [
                     'endpoint' => 'https://codedeploy-fips.us-gov-east-1.amazonaws.com',
                     'signRegion' => 'us-gov-east-1',
                     'signService' => 'codedeploy',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'us-gov-west-1-fips':
                 return [
                     'endpoint' => 'https://codedeploy-fips.us-gov-west-1.amazonaws.com',
                     'signRegion' => 'us-gov-west-1',
                     'signService' => 'codedeploy',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'us-west-1-fips':
                 return [
                     'endpoint' => 'https://codedeploy-fips.us-west-1.amazonaws.com',
                     'signRegion' => 'us-west-1',
                     'signService' => 'codedeploy',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'us-west-2-fips':
                 return [
                     'endpoint' => 'https://codedeploy-fips.us-west-2.amazonaws.com',
                     'signRegion' => 'us-west-2',
                     'signService' => 'codedeploy',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
         }
 
         throw new UnsupportedRegion(sprintf('The region "%s" is not supported by "CodeDeploy".', $region));
-    }
-
-    protected function getServiceCode(): string
-    {
-        return 'codedeploy';
-    }
-
-    protected function getSignatureScopeName(): string
-    {
-        return 'codedeploy';
-    }
-
-    protected function getSignatureVersion(): string
-    {
-        return 'v4';
     }
 }