+++ src/Service/CognitoIdentityProvider/src/CognitoIdentityProviderClient.php
@@ -242,57 +242,34 @@ class CognitoIdentityProviderClient extends AbstractApi
             case 'us-east-2':
             case 'us-west-2':
                 return [
-                    'endpoint' => 'https://cognito-idp.%region%.amazonaws.com',
+                    'endpoint' => "https://cognito-idp.$region.amazonaws.com",
                     'signRegion' => $region,
                     'signService' => 'cognito-idp',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'fips-us-east-1':
                 return [
                     'endpoint' => 'https://cognito-idp-fips.us-east-1.amazonaws.com',
                     'signRegion' => 'us-east-1',
                     'signService' => 'cognito-idp',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'fips-us-east-2':
                 return [
                     'endpoint' => 'https://cognito-idp-fips.us-east-2.amazonaws.com',
                     'signRegion' => 'us-east-2',
                     'signService' => 'cognito-idp',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'fips-us-west-2':
                 return [
                     'endpoint' => 'https://cognito-idp-fips.us-west-2.amazonaws.com',
                     'signRegion' => 'us-west-2',
                     'signService' => 'cognito-idp',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
         }
 
         throw new UnsupportedRegion(sprintf('The region "%s" is not supported by "CognitoIdentityProvider".', $region));
-    }
-
-    protected function getServiceCode(): string
-    {
-        return 'cognito-idp';
-    }
-
-    protected function getSignatureScopeName(): string
-    {
-        return 'cognito-idp';
-    }
-
-    protected function getSignatureVersion(): string
-    {
-        return 'v4';
     }
 }