+++ src/Service/DynamoDb/src/DynamoDbClient.php
@@ -428,140 +428,99 @@ class DynamoDbClient extends AbstractApi
             case 'us-west-1':
             case 'us-west-2':
                 return [
-                    'endpoint' => 'https://dynamodb.%region%.amazonaws.com',
+                    'endpoint' => "https://dynamodb.$region.amazonaws.com",
                     'signRegion' => $region,
                     'signService' => 'dynamodb',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'cn-north-1':
             case 'cn-northwest-1':
                 return [
-                    'endpoint' => 'https://dynamodb.%region%.amazonaws.com.cn',
+                    'endpoint' => "https://dynamodb.$region.amazonaws.com.cn",
                     'signRegion' => $region,
                     'signService' => 'dynamodb',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'us-gov-east-1':
             case 'us-gov-west-1':
                 return [
-                    'endpoint' => 'https://dynamodb.%region%.amazonaws.com',
+                    'endpoint' => "https://dynamodb.$region.amazonaws.com",
                     'signRegion' => $region,
                     'signService' => 'dynamodb',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'us-isob-east-1':
                 return [
-                    'endpoint' => 'https://dynamodb.%region%.sc2s.sgov.gov',
+                    'endpoint' => "https://dynamodb.$region.sc2s.sgov.gov",
                     'signRegion' => $region,
                     'signService' => 'dynamodb',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'ca-central-1-fips':
                 return [
                     'endpoint' => 'https://dynamodb-fips.ca-central-1.amazonaws.com',
                     'signRegion' => 'ca-central-1',
                     'signService' => 'dynamodb',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'local':
                 return [
                     'endpoint' => 'http://localhost:8000',
                     'signRegion' => 'us-east-1',
                     'signService' => 'dynamodb',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'us-east-1-fips':
                 return [
                     'endpoint' => 'https://dynamodb-fips.us-east-1.amazonaws.com',
                     'signRegion' => 'us-east-1',
                     'signService' => 'dynamodb',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'us-east-2-fips':
                 return [
                     'endpoint' => 'https://dynamodb-fips.us-east-2.amazonaws.com',
                     'signRegion' => 'us-east-2',
                     'signService' => 'dynamodb',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'us-gov-east-1-fips':
                 return [
                     'endpoint' => 'https://dynamodb.us-gov-east-1.amazonaws.com',
                     'signRegion' => 'us-gov-east-1',
                     'signService' => 'dynamodb',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'us-gov-west-1-fips':
                 return [
                     'endpoint' => 'https://dynamodb.us-gov-west-1.amazonaws.com',
                     'signRegion' => 'us-gov-west-1',
                     'signService' => 'dynamodb',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'us-iso-east-1':
                 return [
                     'endpoint' => 'https://dynamodb.us-iso-east-1.c2s.ic.gov',
                     'signRegion' => 'us-iso-east-1',
                     'signService' => 'dynamodb',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'us-west-1-fips':
                 return [
                     'endpoint' => 'https://dynamodb-fips.us-west-1.amazonaws.com',
                     'signRegion' => 'us-west-1',
                     'signService' => 'dynamodb',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'us-west-2-fips':
                 return [
                     'endpoint' => 'https://dynamodb-fips.us-west-2.amazonaws.com',
                     'signRegion' => 'us-west-2',
                     'signService' => 'dynamodb',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
         }
 
         throw new UnsupportedRegion(sprintf('The region "%s" is not supported by "DynamoDb".', $region));
-    }
-
-    protected function getServiceCode(): string
-    {
-        return 'dynamodb';
-    }
-
-    protected function getSignatureScopeName(): string
-    {
-        return 'dynamodb';
-    }
-
-    protected function getSignatureVersion(): string
-    {
-        return 'v4';
     }
 }