+++ src/Service/Lambda/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## NOT RELEASED
 
+### Deprecation
+
+- Removes methods `getServiceCode`, `getSignatureVersion` and `getSignatureScopeName` from Client.
+
 ### Fixed
 
 - Fixed issue when Layer, Function or Version contained a special char `#`